pois e

+++ src/screen_logo_cin.cpp
@@ -28,7 +28,7 @@ void LogoCINScreen::Init() {
     revealSpeed = 2.0f;
 
     cin_image = LoadImage("rsc\\HC.png");
-    ImageResize(&cin_image, scale * cin_image.width, scale * cin_image.height);
+    ImageResize(&cin_image, (int) scale * cin_image.width, (int) scale * cin_image.height);
     texture = LoadTextureFromImage(cin_image);
 }
